made addTask const a seperate component for clean code
